backend/requirements.txt
@@ .. @@
 # Core Flask dependencies
 Flask==2.3.3
 Flask-CORS==4.0.0
 python-dotenv==1.0.0
 
 # AI/ML Libraries
 transformers==4.35.2
 torch==2.1.0
 huggingface-hub==0.19.4
 openai==1.3.5
 
 # Text Processing
 spacy==3.7.2
 langdetect==1.0.9
 newspaper3k==0.2.8
 
 # Data Processing
 pandas==2.1.3
 numpy==1.24.3
 requests==2.31.0
 
 # RSS/Feed Parsing
 feedparser==6.0.10
 
 # Google APIs
 google-cloud-translate==3.12.1
 google-api-python-client==2.108.0
 
 # Web scraping utilities
 beautifulsoup4==4.12.2
 lxml==4.9.3
 
 # Additional utilities
 dateparser==1.2.0
-validators==0.22.0
+validators==0.22.0
+python-dateutil==2.8.2
+gunicorn==21.2.0